// src/types/FormResponseToAPI.ts
export interface FormResponse {
  foretakTelefonnummer: string;
  foretakWebPageUrl: string;
  norskForetaksNavn: string;
  harFilialINorge: boolean;
  filialAdresse: string;
  filialPostnummer: string;
  filialPoststed: string;
  filialTelefonnummer: string;
  filialWebPageUrl: string;
  filialNaceKode1: string;
  filialNaceKode2: string;
  filialNaceKode3: string;
}